+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- disable save button when there is no editor
+
 ## [0.1.10] - 2022-05-25
 ### Changed
 - removed some extraneous toolbar buttons